requirements.txt
streamlit
pandas
psutil
mysql-connector-python
python-dotenv
requests

